IMM精度

+++ IMM-CalibrateTool/IMM-CalibrateTool/common/cfgcom/config.h
@@ -30,7 +30,7 @@ struct sErrData
 {
     ushort vol;
     ushort volErr;
-    ushort curErr;
+    double curErr;
     ushort powErr; // 电流误差
 };
